+++ Source/Core/InputCommon/DynamicInputTextures/DITSpecification.h
@@ -1,0 +1,18 @@
+// Copyright 2021 Dolphin Emulator Project
+// Licensed under GPLv2+
+// Refer to the license.txt file included.
+
+#pragma once
+
+#include <string>
+#include <vector>
+
+#include <picojson.h>
+
+#include "InputCommon/DynamicInputTextures/DITData.h"
+
+namespace InputCommon::DynamicInputTextures
+{
+bool ProcessSpecificationV1(picojson::value& root, std::vector<Data>& input_textures,
+                            const std::string& base_path, const std::string& json_file);
+}
